git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3861 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/basicimageimporters/qfimagereaderraw.h
@@ -56,7 +56,7 @@ class QFImageReaderRAW: public QFImporterImageSeries
                 width=set.value("width", 0).toInt();
                 height=set.value("height", 0).toInt();
                 channels=set.value("channels", 1).toInt();
-                int bps=set.value("bitspersample", sizeof(T)*8).toInt();
+                int bps=set.value("bitspersample", int(sizeof(T)*8)).toInt();
                 ok=((bps==sizeof(T)*8)&&(width>0)&&(height>0)&&(channels>0));
             } else if (QFile::exists(filename)) {
                 ok=true;
